# Reduce `ScannerQueueTime` in HdfsScanNode