@dirtysalt 2022-03-22
There is no any evidence that this PR can improve HDFSScanNode active time, and there is a risk that some scanner will execute more time(higher priority), which leads to un-fairness. So better to close this PR.